Add plain option to renderEntryContent.
Default false. When set to true, this returns blocks that output plain markup without the Marquee classes added, or additional information such as that used by ImageBlocks and EmbedBlocks. This is useful for cleaner RSS feeds and other contexts that don’t support CSS and/or JavaScript. (ImageBlocks normally need JS to render themselves responsively client-side.)